Separating duties a bit more. There's a lag between downloading a package and retrieving the next data from the package listing. This causes the package listing to timeout. Need a paralell download structure.
